cleaning up api for fitness, genotype, simulation, engine
