+++ eee/evolve/engine/wright_fisher.py
@@ -2,7 +2,7 @@
 Run a Wright-Fisher simulation given an ensemble.
 """
 
-from eee.evolve.genotype import GenotypeContainer
+from eee.evolve.genotype.genotype_container import GenotypeContainer
 from eee.evolve.analysis import get_num_accumulated_mutations
 from eee._private.interface import MockTqdm
