+++ eee/evolve/fitness/ff.py
@@ -6,21 +6,26 @@ may also take further keyword arguments (passed in via fitness_kwargs). Whatever
 occurs after ff_ will be the name of the function in things like json files 
 using the function. 
 
-def ff_example_function(value,multiply_by):
-    '''
-    Return the fitness of a value as that value times multiply_by. 
-    '''
+.. code-block::python
 
-    return value*multiply_by
+    def ff_example_function(value,multiply_by):
+        '''
+        Return the fitness of a value as that value times multiply_by. 
+        '''
+
+        return value*multiply_by
 
 We could use this in a simulation set up as follows. This set of kwargs will
 apply the function ff_example_function to "fx_obs." ff_example_function will 
 multiply "fx_obs" by 5 at an iptg chemical potential of 1. 
 
-{"fitness_fcns":["example_function"],
- "select_on": "fx_obs",
- "fitness_kwargs":{"multiply_by":5},
- "mu_dict":{"iptg":1}}
+.. code-block::python
+
+    {"fitness_fcns":["example_function"],
+    "select_on": "fx_obs",
+    "fitness_kwargs":{"multiply_by":5},
+    "mu_dict":{"iptg":1}}
+    
 """
 
 def ff_on(value):
