+++ eee/evolve/genotype/__init__.py
@@ -1,0 +1,2 @@
+
+from .genotype_container import GenotypeContainer
